+++ packages/slate-plugins/src/elements/code-block/components/index.ts
@@ -1,0 +1,2 @@
+export * from 'elements/code-block/components/CodeBlockElement';
+export * from 'elements/code-block/components/ToolbarCodeBlock';