+++ packages/slate-plugins/src/elements/code-block/deserializeCodeBlock.ts
@@ -1,0 +1,9 @@
+import { DeserializeHtml } from 'common/types';
+import { getElementDeserializer } from 'element/utils';
+import { CODE_BLOCK, CodeBlockDeserializeOptions } from './types';
+
+export const deserializeCodeBlock = ({
+  typeCodeBlock = CODE_BLOCK,
+}: CodeBlockDeserializeOptions = {}): DeserializeHtml => ({
+  element: getElementDeserializer(typeCodeBlock, { tagNames: ['PRE'] }),
+});