+++ packages/slate-plugins/src/elements/code-block/types.ts
@@ -1,0 +1,40 @@
+import { RenderElementOptions } from 'element';
+import { Element } from 'slate';
+import { RenderElementProps } from 'slate-react';
+
+export const CODE_BLOCK = 'code_block';
+
+// Data of Element node
+export interface CodeBlockNodeData {}
+
+// Element node
+export interface CodeBlockNode extends Element, CodeBlockNodeData {}
+
+// Option type
+interface TypeOption {
+  typeCodeBlock?: string;
+}
+
+// deserialize options
+export interface CodeBlockDeserializeOptions extends TypeOption {}
+
+// renderElement options given as props
+interface CodeBlockRenderElementOptionsProps {}
+
+// renderElement options
+export interface CodeBlockRenderElementOptions
+  extends RenderElementOptions,
+    CodeBlockRenderElementOptionsProps,
+    TypeOption {}
+
+// renderElement props
+export interface CodeBlockRenderElementProps
+  extends RenderElementProps,
+    CodeBlockRenderElementOptionsProps {
+  element: CodeBlockNode;
+}
+
+// Plugin options
+export interface CodeBlockPluginOptions
+  extends CodeBlockRenderElementOptions,
+    CodeBlockDeserializeOptions {}