+++ packages/slate-plugins/src/elements/code/deserializeCode.ts
@@ -1,9 +1,0 @@
-import { DeserializeHtml } from 'common/types';
-import { getElementDeserializer } from 'element/utils';
-import { CODE, CodeDeserializeOptions } from './types';
-
-export const deserializeCode = ({
-  typeCode = CODE,
-}: CodeDeserializeOptions = {}): DeserializeHtml => ({
-  element: getElementDeserializer(typeCode, { tagNames: ['PRE'] }),
-});