+++ packages/slate-plugins/src/marks/code/deserializeCode.ts
@@ -1,0 +1,11 @@
+import { DeserializeHtml } from 'common/types';
+import { getLeafDeserializer } from 'mark/utils';
+import { CodeDeserializeOptions, MARK_CODE } from './types';
+
+export const deserializeCode = ({
+  typeCode = MARK_CODE,
+}: CodeDeserializeOptions = {}): DeserializeHtml => ({
+  leaf: getLeafDeserializer(typeCode, {
+    tagNames: ['CODE', 'KBD'],
+  }),
+});